succesffuly impleted deleting transactions
Implemented this by creating a new column in the table that stores the
key/id of the object in Firebase. This column is hidden, and when a row is
selected to be deleted, the key/id is used to delete the object in Firebase.
The table updates automatically.

Next, I am considering making it so only the oldest transaction can be
deleted.